+++ YTVimeoExtractor/YTVimeoExtractor.m
@@ -26,8 +26,11 @@
     return sharedExtractor;
 }
 
-- (id)init {
-    if (self = [super init]) {
+- (instancetype)init {
+    
+    self = [super init];
+    
+    if (self) {
         
         _extractorOperationQueue = [[NSOperationQueue alloc]init];
         if ([_extractorOperationQueue respondsToSelector:@selector(qualityOfService)]) {