+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+
+### Version 1.2.0
+
+* New `HTTPLiveStreamURL` property on the `YTVimeoVideo` class. ([#38](https://github.com/lilfaf/YTVimeoExtractor/issues/38))
+* Add `YTVimeoVideoQualityMedium540` enum value on the `YTVimeoVideo` class. ([#40](https://github.com/lilfaf/YTVimeoExtractor/issues/40))
+
 ### Version 1.1.0
 
 * The `streamURLs` and `thumbnailURLs` are now dictionaries that contain NSURLs ([e179efc](https://github.com/lilfaf/YTVimeoExtractor/commit/e179efc395ea8f287a9a9acb1b4b1398e92e24ce), [385ae37](https://github.com/lilfaf/YTVimeoExtractor/commit/385ae372660f52a0061b3c73f7d5df0cd5b33cc3))